OS class for real-time censoring
